+++ Examples/Nodify.Playground/Editor/NodifyEditorViewModel.cs
@@ -9,13 +9,13 @@ namespace Nodify.Playground
         public NodifyEditorViewModel()
         {
             Schema = new GraphSchema();
-            
+
             PendingConnection = new PendingConnectionViewModel
             {
                 Graph = this
             };
 
-            DeleteSelectionCommand = new DelegateCommand(DeleteSelection, () => SelectedNodes.Count > 0);
+            DeleteSelectionCommand = new DelegateCommand(DeleteSelection, () => SelectedNodes.Count > 0 || SelectedConnections.Count > 0);
             CommentSelectionCommand = new RequeryCommand(() => Schema.AddCommentAroundNodes(SelectedNodes, "New comment"), () => SelectedNodes.Count > 0);
             DisconnectConnectorCommand = new DelegateCommand<ConnectorViewModel>(c => c.Disconnect());
             CreateConnectionCommand = new DelegateCommand<object>(target => Schema.TryAddConnection(PendingConnection.Source!, target), target => PendingConnection.Source != null && target != null);
@@ -63,6 +63,13 @@ namespace Nodify.Playground
             set => SetProperty(ref _selectedNodes, value);
         }
 
+        private NodifyObservableCollection<ConnectionViewModel> _selectedConnections = new NodifyObservableCollection<ConnectionViewModel>();
+        public NodifyObservableCollection<ConnectionViewModel> SelectedConnections
+        {
+            get => _selectedConnections;
+            set => SetProperty(ref _selectedConnections, value);
+        }
+
         private NodifyObservableCollection<ConnectionViewModel> _connections = new NodifyObservableCollection<ConnectionViewModel>();
         public NodifyObservableCollection<ConnectionViewModel> Connections
         {
@@ -78,6 +85,21 @@ namespace Nodify.Playground
         }
 
         public PendingConnectionViewModel PendingConnection { get; }
+
+        private ConnectionViewModel? _selectedConnection;
+        public ConnectionViewModel? SelectedConnection
+        {
+            get => _selectedConnection;
+            set => SetProperty(ref _selectedConnection, value);
+        }
+
+        private NodeViewModel? _selectedNode;
+        public NodeViewModel? SelectedNode
+        {
+            get => _selectedNode;
+            set => SetProperty(ref _selectedNode, value);
+        }
+
         public GraphSchema Schema { get; }
 
         public ICommand DeleteSelectionCommand { get; }
@@ -87,6 +109,11 @@ namespace Nodify.Playground
 
         private void DeleteSelection()
         {
+            foreach (var connection in SelectedConnections.ToList())
+            {
+                connection.Remove();
+            }
+
             var selected = SelectedNodes.ToList();
 
             for (int i = 0; i < selected.Count; i++)