+++ Examples/Nodify.Playground/EditorSettings.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Collections.ObjectModel;
 using System.Windows;
 
 namespace Nodify.Playground
@@ -14,18 +13,37 @@ namespace Nodify.Playground
 
     public class EditorSettings : ObservableObject
     {
-        public IReadOnlyCollection<ISettingViewModel> Settings { get; }
-        public IReadOnlyCollection<ISettingViewModel> AdvancedSettings { get; }
+        private readonly IReadOnlyCollection<ISettingViewModel> _settings;
+        public IEnumerable<ISettingViewModel> Settings => PlaygroundSettings.Instance.FilterAndSort(_settings);
+
+        private readonly IReadOnlyCollection<ISettingViewModel> _advancedSettings;
+        public IEnumerable<ISettingViewModel> AdvancedSettings => PlaygroundSettings.Instance.FilterAndSort(_advancedSettings);
 
         private EditorSettings()
         {
-            Settings = new ObservableCollection<ISettingViewModel>()
+            PlaygroundSettings.Instance.PropertyChanged += OnSearchTextChanged;
+
+            _settings = new List<ISettingViewModel>()
             {
                 new ProxySettingViewModel<bool>(
                     () => Instance.EnableRealtimeSelection,
                     val => Instance.EnableRealtimeSelection = val,
                     "Realtime selection: ",
                     "Selects items when finished if disabled."),
+                new ProxySettingViewModel<bool>(
+                    () => Instance.SelectableNodes,
+                    val => Instance.SelectableNodes = val,
+                    "Selectable nodes: ",
+                    "Whether nodes can be selected."),
+                new ProxySettingViewModel<bool>(
+                    () => Instance.DraggableNodes,
+                    val => Instance.DraggableNodes= val,
+                    "Draggable nodes: ",
+                    "Whether nodes can be dragged."),
+                new ProxySettingViewModel<bool>(
+                    () => Instance.CanSelectMultipleNodes,
+                    val => Instance.CanSelectMultipleNodes = val,
+                    "Can select multiple nodes: "),
                 new ProxySettingViewModel<bool>(
                     () => Instance.EnablePendingConnectionSnapping,
                     val => Instance.EnablePendingConnectionSnapping = val,
@@ -86,6 +104,15 @@ namespace Nodify.Playground
                     val => Instance.AutoPanningEdgeDistance = val,
                     "Auto panning edge distance: ",
                     "Distance from edge to trigger auto panning"),
+                new ProxySettingViewModel<bool>(
+                    () => Instance.SelectableConnections,
+                    val => Instance.SelectableConnections = val,
+                    "Selectable connections: ",
+                    "Whether connections can be selected."),
+                new ProxySettingViewModel<bool>(
+                    () => Instance.CanSelectMultipleConnections,
+                    val => Instance.CanSelectMultipleConnections = val,
+                    "Can select multiple connections: "),
                 new ProxySettingViewModel<ConnectionStyle>(
                     () => Instance.ConnectionStyle,
                     val => Instance.ConnectionStyle = val,
@@ -99,6 +126,11 @@ namespace Nodify.Playground
                     val => Instance.CircuitConnectionAngle = val,
                     "Connection angle: ",
                     "Applies to circuit connection style"),
+                new ProxySettingViewModel<double>(
+                    () => Instance.ConnectionCornerRadius,
+                    val => Instance.ConnectionCornerRadius = val,
+                    "Connection corner radius: ",
+                    "The corner radius between the line segments."),
                 new ProxySettingViewModel<double>(
                     () => Instance.ConnectionSpacing,
                     val => Instance.ConnectionSpacing = val,
@@ -119,6 +151,16 @@ namespace Nodify.Playground
                     val => Instance.DirectionalArrowsOffset = val,
                     "Directional arrows offset: ",
                     "Used to animate the directional arrowheads flowing in the direction of the connection (value is between 0 and 1)."),
+                new ProxySettingViewModel<bool>(
+                    () => Instance.IsAnimatingConnections,
+                    val => Instance.IsAnimatingConnections = val,
+                    "Animate directional arrows: ",
+                    "Used to animate the directional arrowheads by animating the DirectionalArrowsOffset value"),
+                new ProxySettingViewModel<double>(
+                    () => Instance.DirectionalArrowsAnimationDuration,
+                    val => Instance.DirectionalArrowsAnimationDuration = val,
+                    "Arrows animation duration: ",
+                    "The duration in seconds of a directional arrowhead flowing from start to end."),
                 new ProxySettingViewModel<ArrowHeadEnds>(
                     () => Instance.ArrowHeadEnds,
                     val => Instance.ArrowHeadEnds = val,
@@ -166,7 +208,7 @@ namespace Nodify.Playground
                     "Whether the grouping node is sticky or not"),
             };
 
-            AdvancedSettings = new ObservableCollection<ISettingViewModel>()
+            _advancedSettings = new List<ISettingViewModel>()
             {
                 new ProxySettingViewModel<double>(
                     () => Instance.HandleRightClickAfterPanningThreshold,
@@ -251,6 +293,15 @@ namespace Nodify.Playground
             };
 
             EnableCuttingLinePreview = true;
+        }
+
+        private void OnSearchTextChanged(object? sender, System.ComponentModel.PropertyChangedEventArgs e)
+        {
+            if (e.PropertyName == nameof(PlaygroundSettings.SearchText))
+            {
+                OnPropertyChanged(nameof(Settings));
+                OnPropertyChanged(nameof(AdvancedSettings));
+            }
         }
 
         public static EditorSettings Instance { get; } = new EditorSettings();
@@ -355,6 +406,41 @@ namespace Nodify.Playground
             set => SetProperty(ref _location, value);
         }
 
+        private bool _selectableConnections = true;
+        public bool SelectableConnections
+        {
+            get => _selectableConnections;
+            set => SetProperty(ref _selectableConnections, value);
+        }
+
+        private bool _canSelectMultipleConnections = true;
+        public bool CanSelectMultipleConnections
+        {
+            get => _canSelectMultipleConnections;
+            set => SetProperty(ref _canSelectMultipleConnections, value);
+        }
+
+        private bool _draggableNodes = true;
+        public bool DraggableNodes
+        {
+            get => _draggableNodes;
+            set => SetProperty(ref _draggableNodes, value);
+        }
+
+        private bool _selectableNodes = true;
+        public bool SelectableNodes
+        {
+            get => _selectableNodes;
+            set => SetProperty(ref _selectableNodes, value);
+        }
+
+        private bool _canSelectMultipleNodes = true;
+        public bool CanSelectMultipleNodes
+        {
+            get => _canSelectMultipleNodes;
+            set => SetProperty(ref _canSelectMultipleNodes, value);
+        }
+
         private ConnectionStyle _connectionStyle;
         public ConnectionStyle ConnectionStyle
         {
@@ -374,6 +460,13 @@ namespace Nodify.Playground
         {
             get => _circuitConnectionAngle;
             set => SetProperty(ref _circuitConnectionAngle, value);
+        }
+
+        private double _connectionCornerRadius = 10;
+        public double ConnectionCornerRadius
+        {
+            get => _connectionCornerRadius;
+            set => SetProperty(ref _connectionCornerRadius, value);
         }
 
         private double _connectionSpacing = 20;
@@ -437,6 +530,20 @@ namespace Nodify.Playground
         {
             get => _directionalArrowsOffset;
             set => SetProperty(ref _directionalArrowsOffset, value);
+        }
+
+        private bool _isAnimatingConnections;
+        public bool IsAnimatingConnections
+        {
+            get => _isAnimatingConnections;
+            set => SetProperty(ref _isAnimatingConnections, value);
+        }
+
+        private double _directionalArrowsAnimationDuration = 2.0;
+        public double DirectionalArrowsAnimationDuration
+        {
+            get => _directionalArrowsAnimationDuration;
+            set => SetProperty(ref _directionalArrowsAnimationDuration, value);
         }
 
         private PointEditor _connectionArrowSize = new Size(8, 8);